Python/Bot/TgOllamaDeepSeek/requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
logging>=0.4.9